+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 #### Breaking Changes:
 * Subject attributes: the `subject` parameter of the assignment function was changed from a string to an object. The new `subject` object contains a `key` field for the subject ID as well as an optional `custom_attributes` property for any related metadata like name or email.
+* The EppoClient `assign()` function is renamed to `get_assignment()`
 
 ## [0.0.3] - 2022-05-11